requirements.txt



Flask==2.3.3
python-dotenv==1.0.0
Pillow==10.0.1
pdfplumber==0.9.0
pandas==2.1.1
google-generativeai==0.3.2
easyocr==1.7.0
numpy==1.24.3
scikit-learn==1.3.0
openpyxl==3.1.2
Werkzeug==2.3.7